move parser index util
