+++ de.bitub.step.p21/src/de/bitub/step/p21/parser/P21LevelListener.java
@@ -2,11 +2,11 @@ package de.bitub.step.p21.parser;
 
 import org.antlr.v4.runtime.ParserRuleContext;
 
-import de.bitub.step.p21.IndexUtil;
 import de.bitub.step.p21.StepParser;
 import de.bitub.step.p21.StepParser.ListContext;
 import de.bitub.step.p21.StepParser.ParameterContext;
 import de.bitub.step.p21.StepParser.ParameterListContext;
+import de.bitub.step.p21.parser.util.IndexUtil;
 import de.bitub.step.p21.StepParserBaseListener;
 import de.bitub.step.p21.StepParserListener;
