+++ de.bitub.step.p21/src/de/bitub/step/p21/parser/util/IndexUtilImpl.java
@@ -1,4 +1,4 @@
-package de.bitub.step.p21;
+package de.bitub.step.p21.parser.util;
 
 import java.util.ArrayDeque;
 import java.util.Deque;
